minor: added webserver config
