+++ .env.example
@@ -5,3 +5,4 @@ LOG_STDOUT=true
 LOG_LEVEL=debug
 SHELL=/usr/bin/bash
 SHELL_ARGS=-c
+TZ=Asia/Tehran
