+++ config/validators.go
@@ -1,14 +1,10 @@
 package config
 
 import (
-	"encoding/json"
-	"errors"
 	"fmt"
 
+	"github.com/google/uuid"
 	"github.com/sirupsen/logrus"
-
-	"github.com/FMotalleb/crontab-go/core/event"
-	credential "github.com/FMotalleb/crontab-go/core/os_credential"
 )
 
 // Validate checks the validity of the Config struct.
@@ -26,6 +22,10 @@ func (cfg *Config) Validate(log *logrus.Entry) error {
 		return err
 	}
 
+	if err := validateWebserverConfig(cfg); err != nil {
+		return err
+	}
+
 	// Validate each job in the config
 	for _, job := range cfg.Jobs {
 		if err := job.Validate(log); err != nil {
@@ -37,238 +37,22 @@ func (cfg *Config) Validate(log *logrus.Entry) error {
 	return nil
 }
 
-// Validate checks the validity of a JobConfig.
-// It ensures that the job is not disabled and all its events, tasks, done hooks, and failed hooks are valid.
-// If any validation fails, it returns an error with the specific validation error.
-// Otherwise, it returns nil.
-func (c *JobConfig) Validate(log *logrus.Entry) error {
-	// Log the start of validation
-	log.Tracef("Validating JobConfig: %s", c.Name)
-
-	// Check if the job is disabled
-	if c.Disabled {
-		// Log the disabled job
-		log.Debugf("JobConfig %s is disabled", c.Name)
+func validateWebserverConfig(cfg *Config) error {
+	if cfg.WebServerAddress == "" {
+		cfg.debugLog("no webserver address specified")
 		return nil
 	}
-	checkList := []func(*JobConfig, *logrus.Entry) error{
-		validateEvents,
-		validateTasks,
-		validateDoneHooks,
-		validateFailedHooks,
-	}
-	for _, check := range checkList {
-		if err := check(c, log); err != nil {
-			return err
-		}
+	if cfg.WebServerAddress != "" && cfg.WebServerPort == 0 {
+		return fmt.Errorf("address: %s:%d is not a valid address", cfg.WebServerAddress, cfg.WebServerPort)
 	}
 
-	// Log the successful validation
-	log.Tracef("Validation successful for JobConfig: %s", c.Name)
-	return nil
-}
-
-func validateFailedHooks(c *JobConfig, log *logrus.Entry) error {
-	for _, t := range c.Hooks.Failed {
-		if err := t.Validate(log); err != nil {
-			log.Errorf("Validation error in failed hook for JobConfig %s: %v", c.Name, err)
-			return err
-		}
-	}
-	return nil
-}
-
-func validateDoneHooks(c *JobConfig, log *logrus.Entry) error {
-	for _, t := range c.Hooks.Done {
-		if err := t.Validate(log); err != nil {
-			log.Errorf("Validation error in done hook for JobConfig %s: %v", c.Name, err)
-			return err
-		}
-	}
-	return nil
-}
-
-func validateTasks(c *JobConfig, log *logrus.Entry) error {
-	for _, t := range c.Tasks {
-		if err := t.Validate(log); err != nil {
-			log.Errorf("Validation error in task for JobConfig %s: %v", c.Name, err)
-			return err
-		}
-	}
-	return nil
-}
-
-func validateEvents(c *JobConfig, log *logrus.Entry) error {
-	for _, s := range c.Events {
-		if err := s.Validate(log); err != nil {
-			log.Errorf("Validation error in event for JobConfig %s: %v", c.Name, err)
-			return err
-		}
-	}
-	return nil
-}
-
-// Validate checks the validity of a Task.
-// It ensures that the task has exactly one of the Get, Post, or Command fields, and validates other fields based on the specified action.
-// If any validation fails, it returns an error with the specific validation error.
-// Otherwise, it returns nil.
-func (t *Task) Validate(log *logrus.Entry) error {
-	// Log the start of validation
-	log.Tracef("Validating Task: %+v", t)
-	checkList := []func(*Task, *logrus.Entry) error{
-		validateActionsList,
-		validateCredential,
-		validateFields,
-		validateGetRequest,
-		validateTimeout,
-		validatePostData,
-		validateRetry,
-	}
-	for _, check := range checkList {
-		if err := check(t, log); err != nil {
-			return err
-		}
-	}
-
-	// Validate hooks
-	for _, task := range append(t.OnDone, t.OnFail...) {
-		if err := task.Validate(log); err != nil {
-			joinedErr := errors.Join(errors.New("hook: failed to validate"), err)
-			log.WithError(joinedErr).Warn("Validation failed for Task")
-			return joinedErr
-		}
-	}
-
-	// Log the successful validation
-	log.Tracef("Validation successful for Task: %+v", t)
-	return nil
-}
-
-func validateRetry(t *Task, log *logrus.Entry) error {
-	if t.RetryDelay < 0 {
-		err := fmt.Errorf(
-			"retry delay for tasks cannot be negative received `%d` for %+v",
-			t.RetryDelay,
-			t,
+	if err := uuid.Validate(cfg.WebServerToken); err != nil {
+		return fmt.Errorf(
+			"webserver token must be a valid UUID token, received value: %s, error: %s",
+			cfg.WebServerToken,
+			err,
 		)
-		log.WithError(err).Warn("Validation failed for Task")
-		return err
-	}
-	return nil
-}
-
-func validatePostData(t *Task, log *logrus.Entry) error {
-	if t.Data != nil {
-		_, err := json.Marshal(t.Data)
-		if err != nil {
-			log.WithError(err).Warn("Validation failed for Task")
-			return err
-		}
-	}
-	return nil
-}
-
-func validateTimeout(t *Task, log *logrus.Entry) error {
-	if t.Timeout < 0 {
-		err := fmt.Errorf(
-			"timeout for tasks cannot be negative received `%d` for %+v",
-			t.Timeout,
-			t,
-		)
-		log.WithError(err).Warn("Validation failed for Task")
-		return err
-	}
-	return nil
-}
-
-func validateGetRequest(t *Task, log *logrus.Entry) error {
-	if t.Get != "" && t.Data != nil {
-		err := fmt.Errorf("GET request cannot have data field, violating GET URI: `%s`", t.Get)
-		log.WithError(err).Warn("Validation failed for Task")
-		return err
-	}
-	return nil
-}
-
-func validateFields(t *Task, log *logrus.Entry) error {
-	if t.Command != "" && (t.Data != nil || t.Headers != nil) {
-		err := fmt.Errorf("command cannot have data or headers field, violating command: `%s`", t.Command)
-		log.WithError(err).Warn("Validation failed for Task")
-		return err
-	}
-	return nil
-}
-
-func validateCredential(t *Task, log *logrus.Entry) error {
-	if err := credential.Validate(log, t.UserName, t.GroupName); err != nil {
-		log.WithError(err).Warn("Be careful when using credentials, in local mode you can't use credentials unless running as root")
-	}
-	return nil
-}
-
-func validateActionsList(t *Task, log *logrus.Entry) error {
-	actions := []bool{
-		t.Get != "",
-		t.Command != "",
-		t.Post != "",
-	}
-	activeActions := 0
-	for _, t := range actions {
-		if t {
-			activeActions++
-		}
-	}
-	if activeActions != 1 {
-		err := fmt.Errorf(
-			"a single task should have one of (Get, Post, Command) fields, received:(Command: `%s`, Get: `%s`, Post: `%s`)",
-			t.Command,
-			t.Get,
-			t.Post,
-		)
-		log.WithError(err).Warn("Validation failed for Task")
-		return err
-	}
-	return nil
-}
-
-// Validate checks the validity of a JobEvent configuration.
-// It ensures that the event has a valid interval or cron expression, and only one of on_init, interval, or cron is set.
-// It returns an error if the validation fails, otherwise, it returns nil.
-func (s *JobEvent) Validate(log *logrus.Entry) error {
-	// Check if the interval is a negative value
-	if s.Interval < 0 {
-		err := fmt.Errorf("received a negative time in interval: `%v`", s.Interval)
-		log.WithError(err).Warn("Validation failed for JobEvent")
-		return err
-	} else if _, err := event.CronParser.Parse(s.Cron); s.Cron != "" && err != nil {
-		log.WithError(err).Warn("Validation failed for JobEvent")
-		return err
 	}
 
-	// Check the active events to ensure only one of on_init, interval, or cron is set
-	events := []bool{
-		s.Interval != 0,
-		s.Cron != "",
-		s.OnInit,
-	}
-	activeEvents := 0
-	for _, t := range events {
-		if t {
-			activeEvents++
-		}
-	}
-	if activeEvents != 1 {
-		err := fmt.Errorf(
-			"a single event must have one of (on_init: true,interval,cron) field, received:(on_init: %t,cron: `%s`, interval: `%s`)",
-			s.OnInit,
-			s.Cron,
-			s.Interval,
-		)
-		log.WithError(err).Warn("Validation failed for JobEvent")
-		return err
-	}
-
-	// Log the successful validation
-	log.Tracef("Validation successful for JobEvent: %+v", s)
 	return nil
 }
